fix: <svg>태그 별도의 컴포넌트로 분리
