+++ src/components/detailLayout.tsx
@@ -8,6 +8,7 @@ import ErrorPage from '@/components/error';
 import {fetchReviewsBySwimmingPoolId} from '@/data/firestore';
 import {PublicSwimmingPool, ReviewData} from '@/lib/types';
 import {Spinner, Button, Link} from '@nextui-org/react';
+import {LocationDotIcon, ArrowIcon, PenIcon} from '@/ui/icon';
 
 function SwimmingPoolDetail({item}: {item: PublicSwimmingPool}) {
   return (
@@ -28,16 +29,7 @@ function SwimmingPoolDetail({item}: {item: PublicSwimmingPool}) {
         <br />
         <br />
         <div className="flex items-center mt-20 mb-3">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            className="text-blue-500 w-5 h-5 mr-2"
-            viewBox="0 0 384 512"
-          >
-            <path
-              fill="#1079ea"
-              d="M215.7 499.2C267 435 384 279.4 384 192C384 86 298 0 192 0S0 86 0 192c0 87.4 117 243 168.3 307.2c12.3 15.3 35.1 15.3 47.4 0zM192 128a64 64 0 1 1 0 128 64 64 0 1 1 0-128z"
-            />
-          </svg>
+          <LocationDotIcon />
           <p className="leading-relaxed">{item.SIGUN_NM}</p>
         </div>
         {(item.IRREGULR_RELYSWIMPL_LENG &&
@@ -84,32 +76,12 @@ function ReviewList({
           className="flex items-center text-blue-500 text-sm hover:text-gray-300"
         >
           <p className="mr-2">방문한 수영장 인증</p>
-          <svg
-            className="inline-flex h-6 w-6"
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-              d="M9 5l7 7-7 7"
-            ></path>
-          </svg>
+          <ArrowIcon />
         </Link>
       </div>
       {reviews.length === 0 ? (
         <div className="flex w-full mx-auto justify-center mb-5 flex-wrap bg-gray-100 text-gray-400 rounded-lg overflow-hidden shadow-md p-4">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 512 512"
-            fill="#cfcfcf"
-            className="w-4 h-5 mr-2"
-          >
-            <path d="M362.7 19.3L314.3 67.7 444.3 197.7l48.4-48.4c25-25 25-65.5 0-90.5L453.3 19.3c-25-25-65.5-25-90.5 0zm-71 71L58.6 323.5c-10.4 10.4-18 23.3-22.2 37.4L1 481.2C-1.5 489.7 .8 498.8 7 505s15.3 8.5 23.7 6.1l120.3-35.4c14.1-4.2 27-11.8 37.4-22.2L421.7 220.3 291.7 90.3z" />
-          </svg>
+          <PenIcon />
           <p>가장 처음으로 리뷰를 작성해주세요!</p>
         </div>
       ) : (
